load_master_idx/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.0
    # via requests
google-api-core==2.24.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.37.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery==3.27.0
    # via edgar (pyproject.toml)
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==2.19.0
    # via edgar (pyproject.toml)
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.68.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.68.1
    # via google-api-core
idna==3.10
    # via requests
packaging==24.2
    # via google-cloud-bigquery
proto-plus==1.25.0
    # via google-api-core
protobuf==5.29.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
urllib3==2.3.0
    # via requests
